[COGF-58] Funcionalidades do anunciante (#60)
* [COGF-58] Funcionalidades do anunciante

Introduces a ProductEditBar component for editing and deleting products, visible only to the product's author. Updates navigation after editing a product to redirect to the product page instead of the profile. Adds ProductEditBar styles, integrates it into ProductContent, and enhances AuthProvider with a noSpinner prop to suppress the loading spinner when desired.

* Mock useRouter from next/navigation in tests

Added a mock for the useRouter hook from next/navigation to the ProductContent component tests. This ensures that router-related functionality can be tested without relying on the actual Next.js router.